+++ kernels/xeon/geometry/instance_intersector4.h
@@ -25,8 +25,8 @@ namespace embree
   {
     struct FastInstanceIntersector4
     {
-      static void intersect(bool4* valid, const Instance* instance, Ray4& ray, size_t item);
-      static void occluded (bool4* valid, const Instance* instance, Ray4& ray, size_t item);
+      static void intersect(bool4* valid, const Instance* instance, Ray4& ray, unsigned int item);
+      static void occluded (bool4* valid, const Instance* instance, Ray4& ray, unsigned int item);
     };
   }
 }
